+++ include/tvm/ir/transform.h
@@ -56,11 +56,11 @@
 #ifndef TVM_IR_TRANSFORM_H_
 #define TVM_IR_TRANSFORM_H_
 
+#include <tvm/ir/diagnostic_context.h>
 #include <tvm/ir/error.h>
 #include <tvm/ir/module.h>
 #include <tvm/node/container.h>
 #include <tvm/runtime/container.h>
-#include <tvm/support/diagnostic_context.h>
 #include <tvm/support/with.h>
 
 #include <string>
