+++ include/tvm/ir/type_relation.h
@@ -25,11 +25,11 @@
 #define TVM_IR_TYPE_RELATION_H_
 
 #include <tvm/ir/attrs.h>
+#include <tvm/ir/diagnostic_context.h>
 #include <tvm/ir/env_func.h>
 #include <tvm/ir/module.h>
 #include <tvm/ir/type.h>
-#include <tvm/support/diagnostic.h>
-#include <tvm/support/diagnostic_context.h>
+#include <tvm/support/logging.h>
 
 namespace tvm {
